[GTK] Test fast/shadow-dom/input-element-in-shadow.html is failing in the bots https://bugs.webkit.org/show_bug.cgi?id=168423
Patch by Carlos Garcia Campos <cgarcia@igalia.com> on 2017-02-16
Reviewed by Ryosuke Niwa.

It's failing only because the result has the spell checker underline, while the the reference image doesn't. I
don't know why the slot is spell checked and the input in the reference isn't, but we could make the test pass
by using real words, since the contents are not important for this test. Use "one", "two" instead of "abc", "def".

* fast/shadow-dom/input-element-in-shadow-expected.html:
* fast/shadow-dom/input-element-in-shadow.html:

Identifier: 185520@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@212536 268f45cc-cd09-0410-ab3c-d52691b4dbfc