BytecodeGenerator should not iterate its m_controlFlowScopeStack using a pointer bump. https://bugs.webkit.org/show_bug.cgi?id=168585
Reviewed by Yusuke Suzuki.

This is because m_controlFlowScopeStack is a SegmentedVector, and entries for
consecutive indices in the vector are not guaranteed to be consecutive in memory
layout.  Instead, we should be using indexing instead.

This issue was detected by the marathon.js test from
https://bugs.webkit.org/show_bug.cgi?id=168580.

* bytecompiler/BytecodeGenerator.cpp:
(JSC::BytecodeGenerator::emitJumpViaFinallyIfNeeded):
(JSC::BytecodeGenerator::emitReturnViaFinallyIfNeeded):



Identifier: 185600@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@212640 268f45cc-cd09-0410-ab3c-d52691b4dbfc