[Modern Media Controls] Enforce a minimum layout height of 50pt for <audio> https://bugs.webkit.org/show_bug.cgi?id=168801 <rdar://problem/30683453>
Patch by Antoine Quint <graouts@apple.com> on 2017-02-23
Reviewed by Eric Carlson.

Source/WebCore:

We enforce a 50pt minimum layout height for <audio> elements. There is no way
for the page author to override this value since the min-height property is set
as !important in a <style> element contained in a ShadowRoot.

* Modules/modern-media-controls/controls/media-controls.css:
(:host(audio)):

LayoutTests:

Update the existing <audio> metrics test to check that we correctly maintain
a 50pt minimum height even if an inline style attribute sets a shorter height
on an <audio> element, and also check that providing a taller height is respected.

* media/modern-media-controls/audio/audio-controls-metrics-expected.txt:
* media/modern-media-controls/audio/audio-controls-metrics.html:

Identifier: 185807@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@212933 268f45cc-cd09-0410-ab3c-d52691b4dbfc