Add a test verifying cache deduplication is not sensitive to SHA1 collision attack https://bugs.webkit.org/show_bug.cgi?id=168774
Reviewed by Sam Weinig.

We use SHA1 for deduplicating disk cache resources. Since a real world SHA1 collision was demonstrated
recently (http://shattered.io/) we can add a test that shows it can't be used for cache poisoning.

There are two protections in the cache code that both individually stop this type of attack:

- When deduplicating the data is verified to be equal by a bytewise comparison.
- SHA1 computations include random salt unique to cache instance.

Commenting out both protections is needed to make this test fail.

* http/tests/cache/disk-cache/resources/shattered-1-nocollision.pdf: Added.
* http/tests/cache/disk-cache/resources/shattered-1.pdf: Added.
* http/tests/cache/disk-cache/resources/shattered-2-nocollision.pdf: Added.
* http/tests/cache/disk-cache/resources/shattered-2.pdf: Added.

Files shattered-1.pdf and shattered-2.pdf differ visually but have the same SHA1.
The nocollision versions for the ref have the same pixels but don't collide.

* http/tests/cache/disk-cache/shattered-deduplication-expected.html: Added.
* http/tests/cache/disk-cache/shattered-deduplication.html: Added.


Identifier: 185823@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@212951 268f45cc-cd09-0410-ab3c-d52691b4dbfc